requirements.txt
flask>=2.0.0
bleach>=4.0.0
markdown>=3.3.0
gunicorn>=20.0.0
Flask-WTF>=1.0.0
geojson
pyyaml
pytest